split into two applications
There now are three layers compared to the previous two:
 - a flask application receiving github webhooks, distributing them to
   the queue
 - a celery queue + workers doing the git handling and calling out to
   the *receiver* running on hydra
 - *NEW* the receiver, to be run on hydra, a simple python application
   without any non-stdlib dependencies

Communication is secured via a PSK and HMAC authenticated messages

This ensures the main nixbot does not have to be run on hydra, only
the minimal receiver, which is to be kept small and not handle any
logic.

This also includes a little assertion to ensure the comment approving
the build is newer than the latest push of the PR to be built, to
mitigate race conditions when approving.